Service holds store typings, might as well do the same for tracker.
